+++ crates/cheatnet/src/runtime_extensions/forge_runtime_extension/contracts_data.rs
@@ -1,5 +1,5 @@
 use super::cheatcodes::declare::get_class_hash;
-use anyhow::Result;
+use anyhow::{Result, anyhow, bail};
 use bimap::BiMap;
 use camino::Utf8PathBuf;
 use conversions::IntoConv;
@@ -72,6 +72,34 @@ impl ContractsData {
             class_hashes,
             selectors,
         })
+    }
+
+    pub fn try_extend(&mut self, other: &Self) -> Result<()> {
+        for (name, contract_data) in &other.contracts {
+            if self.contracts.contains_key(name) {
+                bail!("duplicate contract name: {name}");
+            }
+
+            if let Some(existing_name) = self.class_hashes.get_by_right(&contract_data.class_hash) {
+                bail!(
+                    "duplicate contract class hash for `{}` and `{}`: {:?}",
+                    existing_name,
+                    name,
+                    contract_data.class_hash
+                );
+            }
+        }
+
+        for (name, contract_data) in &other.contracts {
+            self.class_hashes
+                .insert_no_overwrite(name.clone(), contract_data.class_hash)
+                .map_err(|_| anyhow!("class hash mapping for {name} should be unique"))?;
+            self.contracts.insert(name.clone(), contract_data.clone());
+        }
+
+        self.selectors.extend(other.selectors.clone());
+
+        Ok(())
     }
 
     #[must_use]