+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Added
 
 - `--launch-debugger` flag that allows launching a test in debug mode using `cairo-debugger` crate. Read more [here](https://foundry-rs.github.io/starknet-foundry/snforge-advanced-features/debugging.html#live-debugging)
+- Predeployed contracts are now supported in debugging traces and backtrace
 
 #### Changed
 
@@ -21,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Fixed
 
 - Validation of Scarb's compiler config now properly takes into account `[cairo]` section and keys without `unstable-` prefix
+- Permitted minter storage value of ETH predeployed contract, has now the same value as on mainnet
 
 ### Cast